feat: lots of changes

+++ WndMain.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Diagnostics;
+using System.Reflection;
 using System.Windows.Forms;
 using System.Xml.Linq;
 
@@ -8,9 +9,17 @@ namespace icns4win
     public partial class WndMain : Form
     {
         private string tempPath = "";
+        private string configPath = Path.Combine(Application.StartupPath, "config.json");
 
         public WndMain()
         {
+            // 初始化配置文件
+            if (!BackendConfig.InitConfig(configPath))
+            {
+                MessageBox.Show("配置文件初始化失败，请检查配置文件格式是否正确\n或是否有读写权限", "错误", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                Application.Exit();
+            }
+
             // 创建临时目录
             try
             {
@@ -28,7 +37,7 @@ namespace icns4win
             }
             catch (Exception ex)
             {
-                MessageBox.Show(ex.Message);
+                MessageBox.Show(ex.Message, "错误", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
             InitializeComponent();
         }
@@ -89,14 +98,17 @@ namespace icns4win
         // 退出应用
         private void closeApp()
         {
-            // 删除临时目录
-            try
+            if (tempPath != null)
             {
-                Directory.Delete(tempPath, true);
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.Message);
+                // 删除临时目录
+                try
+                {
+                    Directory.Delete(tempPath, true);
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.Message, "错误", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
             }
         }
 
@@ -118,6 +130,32 @@ namespace icns4win
             WndConfig wndConfig = new WndConfig();
             wndConfig.Owner = this;
             wndConfig.ShowDialog();
+        }
+
+        // 关于
+        private void aboutToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            TaskDialogPage page = new TaskDialogPage()
+            {
+                Text = "Apple Icon Image Format 在 Windows 下的处理工具\n" +
+                $"版本: {Assembly.GetExecutingAssembly().GetName().Version}\n" +
+                "该用户界面由 What_Damon 开发\n" +
+                "使用 MIT 许可证开源\n" +
+                $".NET版本: {Environment.Version.ToString()}",
+                Heading = "关于 icns4win",
+                Caption = "关于",
+                Icon = TaskDialogIcon.Information,
+                DefaultButton = TaskDialogButton.OK,
+                Buttons = { TaskDialogButton.OK }
+            };
+
+            TaskDialogButton result = TaskDialog.ShowDialog(this, page);
+        }
+
+        // 按下 GitHub 仓库按钮
+        private void gitHubToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            Process.Start("explorer.exe", "https://github.com/WhatDamon/icns4win");
         }
     }
 }
